requirements.txt
jupyter>=1.0.0
K3D>=2.9.4
matplotlib>=3.4.1
trimesh>=3.9.14
pytorch-lightning>=1.2.8
scikit-image>=0.18.1
pyrender>=0.1.43
moviepy>=1.0.3
pillow>=7.2.0
tqdm>=4.48.2
scipy>=1.4.1